requirements.txt
certifi==2017.11.5
chardet==3.0.4
decorator==4.1.2
docutils==0.14
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
Kivy-Garden==0.1.4
numpy==1.13.3
opencv-python==3.3.0.10
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pygame==1.9.3
Pygments==2.2.0
PyYAML==3.12
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
